HACK OR SNOOZE API
Hacker News is a tech news aggregator founded by Paul Graham, the man behind Y Combinator. It is one of the best sources online for intellectually stimulating news, and its content is not just limited to software technology. In this project, you'll play mad scientist and make a functioning clone of Hacker News. MWA HA HA HA! IT'S ALIVE! 

The link below will take you to Rithm School's website, where you'll find detailed instructions for this Hacker News project. Be sure to follow each step of the instructions carefully and closely. This project will incorporate everything that you've been learning in this course, which is a lot, so refer back to the relevant sections if you ever get stuck. Don't hesitate to reach out to your TA or mentor for advice along the way.